chatbot/requirements.txt
flask==3.0.2
werkzeug==2.0.2
python-dotenv>=1.0.0
google-generativeai==0.3.2
gunicorn==20.1.0
requests==2.31.0
tenacity==8.2.3